FIXED: Compilation on VS 2003.Net by Ian. Thanks for this fix
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@806 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfTTFWriter.cpp
@@ -38,7 +38,7 @@ inline double PdfLog2(double x) { return log2(x); }
 inline double PdfExp2(double x) { return exp2(x); }
 #else
 inline double PdfLog2(double x) { return log(x)/log(2.0f); }
-inline double PdfExp2(double x) { return pow(2,x); }
+inline double PdfExp2(double x) { return pow(static_cast<double>(2.0f),static_cast<double>(x)); }
 #endif
 }
